python/requirements.txt
openai>=2.6.1
acontext==0.0.1.dev7
python-dotenv>=1.2.1

